yaml2markdown/requirements.txt
PyYAML>=6.0
click>=8.0
Pillow>=9.0
requests>=2.28